Use Manual vs ManualRecord throughout ManualRelocator
Operating on instances of `Manual` vs `ManualRecord` is desirable,
because the latter should just be an implementation detail of the
former.

Note that `Manual.find_by_slug!` is now responsible for raising
exceptions when a `Manual` cannot be found or when multiple `Manuals`
are found with the same slug. Thus the relevant code and specs can be
safely removed.