Inline Manual.build into Manual#initialize
Note that this introduces a small change in behaviour to `Manual` which
will be apparent from the change I've had to make to the `Manual` spec.

Previously if `Manual.new` was called without an ID in the attribute
Hash, then a `KeyError` would be raised when the `:id` key was not
found. Now if no ID is supplied, one will be generated.

Eventually we're hoping to convert `Manual` into a `Mongoid::Document`
and this new behaviour feels closer to the way the latter behaves, so
given that all the other specs pass, I'm happy to make this change.

Note that I've used `Manual.new` directly rather than
`FactoryGirl.build(:manual)` to make it clearer that no ID is being
supplied. However, this then means that I have to supply a title to
avoid a `KeyError` being raised.