Default Manual#title to empty string to avoid KeyError
Since `Manual` has a validation for the presence of `title`, I think
this is safe to do and it has the advantage of not raising a `KeyError`
if no `slug` *or* `title` is provided.

Note that this default will only be used if no `:title` key is provided
in the `attributes` Hash. I'm pretty confident that in most cases this
key *will* be present, even if its value is `nil` or an empty String.

Making this change means we no longer need to provide a default value
for `title` in `ManualsController#new` or
`Manual::PreviewService#ephemeral_manual`, i.e. we are effectively
reducing duplication.

It also means we can improve one of the examples in the `Manual` spec to
make its intention clearer; it was only supplying a `title` to avoid the
`KeyError` exception which is no longer a concern.