Upgrade to Rails 5.0.3
We tried to upgrade the minimum number of gems at one time to allow
Rails to be upgraded to the latest patch release in the 5.0 series.

    bundle update rails actionview railties actionmailer \
    activesupport actionpack rails-dom-testing rack rack-test \
    activemodel sinatra mongoid activemodel tilt --conservative

To do this we also had to relax the constraints on Mongoid and
Sinatra. Mongoid 6 is required for Rails 5. Rails 5 needs Rack 2
and Sinatra 2.0 is the first release of Sinatra that works with Rack 2.