Update generated files
We ran `rake app:update`, inspected the changes and incorporated those
that made sense.

In the case of
`Rails.application.config.action_dispatch.cookies_serializer` we were
previously using the default `:marshal` serializer. With rails 5 the
default has changed to `:json`. In our case we cannot accept this
default as there are still valid cookies stored on client machines in
the `:marshal` format. We therefore use the `:hybrid` format which
will migrate the cookies to the new format when the users log in[1]. At
some point in the future we should be able to switch to the `:json`
format.

I've added #1144 to track this.

[1] http://guides.rubyonrails.org/upgrading_ruby_on_rails.html#cookies-serializer