Avoid persisting updated section when saving to Publishing API fails
This fixes a specific problem when updating a section and giving it the
same title as another section in the same manual. The changed section
would be saved to the local database but would fail to save to
Publishing API because the slug of the section would already be in use.

Note that I'm continuing to mark the manual as draft before sending it
to the publishing api. It's not completely obvious whether it's
necessary to do this before saving the manual and new section to the
publishing api (all tests continue to pass if I set it as draft after
it's been saved to the publishing api) but it feels somewhat safer to
retain this aspect of the service's behaviour.

The user will continue to see an error page when attempting to save a
section with a duplicate title. A further improvement will be to either
avoid trying to save the section in the first place (i.e. because we've
determined that the title is already in use) or to handle the validation
errors we get back from the Publication API.

This should get us closer to being able to remove the following classes
that detect different sections sharing the same slug (i.e. the same
title):

* DuplicateDocumentFinder
* DuplicateDraftDeleter
* MarkedSectionDeleter