Add specs for Manual::CreateService
I've ended up using mocks to test the interaction between the various
objects manipulated in `Manual::CreateService`. I started by using a
real user and manual but struggled to write tests that described the
behaviour.

These specs should make it easier to change the behaviour of this class
in subsequent commits.