Avoid persisting new manual when saving to Publishing API fails
This fixes a specific problem when creating a manual and giving it the
same title as another manual. The new manual would be saved to the
local database but would fail to save to Publishing API because the slug
of the manual would already be in use.

The user will continue to see an error page when attempting to save a
manual with a duplicate title. A further improvement will be to either
avoid trying to save the manual in the first place (i.e. because we've
determined that the title is already in use) or to handle the validation
errors we get back from the Publication API.