Remove explicit dependency on rake gem
It looks as if a constraint on the version of `rake` was introduced in
this commit [1] in order to ensure compatibility with Rails v3.2.

However, this constraint was removed in this later commit [2] when
Rails had been upgraded to v4.

As far as I can see this gem doesn't appear to be used directly in the
application code and so I'm happy to remove this explicit dependency.

[1]: beedd8cea71cd71946c0f6574394fe79840c0f05
[2]: beedd8cea71cd71946c0f6574394fe79840c0f05