Remove redundant test-unit gem
This gem was added in the upgrade for Ruby v2.2 [1], because the version
of Rails in use at the time (v3.2.22.3) needed it. However, now that we
have upgraded to Rails v5.0.3, it no longer seems to be required.

Given the comment in the `Gemfile`, I've also confirmed that I can
start a Rails console in both development & production environments.

[1]: a0aa4406d156ac1f28c9bc4a1f026f28d262c8c4