Pass previous and latest editions to Section constructor
This makes the existing behaviour more explicit. Outside of the tests we
only initialized `Section`s from `Section.find` and
`Manual#build_section`.  `Manual#build_section` constructed a `Section`
without any editions while `Manual.find` constructed a `Section` with
the two most recent editions.

Note the call to `compact` for the array of editions in
`Section#initialize`. This avoids nil elements in the array when
`previous_edition` and/or `latest_edition` is nil. The call to `uniq`
avoids a theoretical problem of the same edition being passed as both
`previous_edition` and `latest_edition`. I've not worried about testing
this behaviour as I'm expecting this to change again the near future.

Note that I'm still storing these editions in the `@editions` array
within `Section` to restrict the size of this commit.

Note the change to the examples for `Section.find` to make it clearer
that it instantiates a Section with the previous and latest editions.

Note the change to the "one published and one withdrawn and one draft"
example in the section_spec. This was unrealistic as it was constructing
a `Section` with 3 editions when they were only ever constructed with 2
at most in production.