Remove Section#editions
I think that using `previous_edition` and `latest_edition` instead of
`editions` makes it easier to see the behaviour of `Section`.

Note that I'm continuing to avoid storing a `previous_edition` and
`latest_edition` in `Section#initialize` when they're both the same
object. I haven't tested this behaviour as I'm hoping/expecting it to go
away in the near future.