Avoid accessing Section#latest_edition in SectionReslugger
The editions are an implementation detail of `Section` and shouldn't be
accessed from outside that class.

By introducing `Section#update_slug!` I can avoid accessing
`Section#latest_edition` from this class.

Note the use of `and_call_original` to restore the original behaviour of
`SlugGenerator.new`. Without this, the stubbing of `SlugGenerator.new`
in the main `before` block was complaining because it was expecting to
be initialized with the slug that's set on the `manual` double for the
rest of the tests.