Inline Manual::UpdateService#export_draft_to_publishing_api
Inlining these methods should make it easier to see the behaviour of the
`#call` method.

This change makes it clear that we're sending a different instance of a
manual to the Publishing API than the one we return from the call to
`#call`. I'm not planning to investigate this any further at the moment.