Improve Section#version_type
I found the use of `Section#has_ever_been_published?` within
`#version_type` confusing and _think_ that introducing and using
`first_edition?` better illustrates the purpose of the method.

Commit 1174e12ca4c8f581f00ec3a04cd68ea1f2d07200 updated the
`has_ever_been_published?` method to take account of the `exported_at`
attibute in the case of there being a single edition. That commit note
explains that the purpose is to ensure we send the correct `update_type`
when saving _new_ sections to the Publishing API. Prior to
1174e12ca4c8f581f00ec3a04cd68ea1f2d07200 it was possible for the first
edition of a section to be saved to the Publishing API with an
`update_type` of `minor`.

Commit 48e2dbc39258c0b0d2dc42349c9f4336163bb44e removes a comment from
the code that explains that the first edition of a Section must be sent
to the Publishing API as a `major` update.