Remove Section#has_ever_been_published?
This is redundant now that it's no longer being used.

I intentionally made this redundant as I found the implementation
confusing. I think part of the problem is the use of `#needs_exporting?`
which checks for the presence of the `exported_at` timestamp. Somewhat
contrary to its name `exported_at` is _only_ set (via
`#mark_as_exported!`) when we publish or unpublish a section from the
Publishing API.

I think we might want to consider renaming `exported_at` but that can
wait for another time.