Introduce SectionViewAdapter#accepts_minor_updates?
We were previously using `Section#has_ever_been_published?` as a proxy
for determining whether a change to a Section could be marked as a minor
update.

Similar to 1dcecbf0ae1f33a39d27c0da5c54140a14e52c08, I think that using
`#first_edition?`, and additionally the state, better describes the
circumstances under which an edition can represent a minor update.

I was hoping that I could rely on `#first_edition?` alone to determine
whether an edition could represent a minor update but that's not
possible because of the way the edit functionality currently works.
Editing a section will _show_ the latest edition and then either update
that edition if it's a draft or create a new edition if it's not. In the
case where we edit the first edition of a Section and that edition has
been published we want the user to be able to mark the change as a minor
update in the knowledge that saving the change will create a new edition
for us. I wonder whether this might all be a bit easier if we somehow
split editing an existing edition from creating a new draft of an
edition.  That's an investigation for another time, though.

Note that I've had to change the order of `symbolize_keys` and `slice`
in `Section#update` so that it correctly handles the attributes from a
new instance of `SectionEdition`. This is because the `#attributes`
method for a new instance of `SectionEdition` only responds to keys as
strings, while a persisted instance responds to keys as strings and
symbols:

```
>> SectionEdition.new.attributes[:_id]
=> nil
>> SectionEdition.new.attributes['_id']
=> BSON::ObjectId('592ed001aeb96e55c8e402fb')

>> SectionEdition.first.attributes[:_id]
=> BSON::ObjectId('53a07717ed915d605b000013')
>> SectionEdition.first.attributes['_id']
=> BSON::ObjectId('53a07717ed915d605b000013')
```

This isn't a problem in the production code because the `latest_edition`
will always be persisted (i.e. `#attributes` responds to keys as
symbols) when we're creating a new draft.