Remove unused Attachment#file method
I found the implementation of this method confusing, because e.g.
the `@attachments` instance variable doesn't appear to be used anywhere
else and the `field` local variable doesn't appear to be defined
anywhere.

It turns out the code is not currently used and I think it's safe to
delete it. I think it may have become redundant in this commit [1].

[1]: 61a27a0917badc6bd1448b342e465030246b5e55