Remove unnecessary guard condition and exception
I can't see how `Services.attachment_api` would ever not return
something and so I think this guard condition is unnecessary.

What's more the raising of a `ApiClientNotPresent` exception is not
covered in the tests and this exception is never explicitly handled.

If it turns out that `Services.attachment_api` does return `nil` for
some reason, we won't be any worse off - we'll just see a
`NoMethodError` instead of the custom exception.