Only unpublish sections which have been previously published
There's a problem when a draft section which hasn't been published and
is then removed, the publishing API will discard the draft section but
the section is also added to Manual#removed_sections which is processed
on publishing the manual, attempting to unpublish an non-existant section
from the publishing API and resulting in a GdsApi::HTTPNotFound error.

So only attempt to unpublish previously published sections, drafts are
already discarded at the removal workflow step.