Add `/lib` to eager_load_paths
The app currently isn't running as the classes in `/lib` are not being
loaded. We've used this approach on other apps e.g. publishing-api but
we could also just put them into `/app` where they probably belong in
any case. This will get the app running while the Rails upgrade is
ongoing.