Add Dockerfile ARG to conditionally precompile assets
The publishing-e2e-tests are being built in production mode now,
and as part of that require the assets to be compiled into the image.

This ARG allows us to only compile the assets when we require allowing
for easy switching back to development mode and not slowing down any
existing usage of this Docker image.