Add "Discard Draft" functionality to the UI
We already had a rake task, `delete_draft_manual` that let us do this as
a dev.  This commit takes the code from that rake task and repurposes it
as a `Manual::DiscardDraftService` and hooks it up to the UI.  It is only
available for manuals that have never been published (this constraint is
the same as the rake task).