Avoid triggering the Naming/VariableNumber cop
This cop doesn't like variable names like foo_1 or foo_1_bar_2, which
is unfortunate because our specs use these a lot.  It doesn't seem to
care about method names, which means all our `let` definitions are fine
it's just our actual variable definitions that trigger it.  We rename
the variables on the lines changed by the previous commit to avoid
triggering the cop, as changing all the foo_1_bar_2 style references
across the specs to keep things consistent would be too painful.