Replace DateTime with Time
This is in response to the Style/DateTime cop which we triggered by
changing some FactoryBot references on lines with DateTime.parse
references on them.  This cop suggests using Time.zone instead of
DateTime so that we preserve time zone information that isn't always
preserved when you use DateTime.  We checked the whole app and fixed
all occurances.  This caused some test failures because some of our
fields were DateTime and in our tests we compare these against Time
instances and the zone information made them no longer compare favourably.
To fix this we changed the field definitions to also use Time, which is
what the fields defined by ::Mongoid::Timestamps do anyway, so it keeps
our time fields and the created_at/updated_at time fields consistent.