Update cucumber.yml to avoid deprecation warnings about tags
Cucumber is moving away from '~@tag' to 'not @tag' syntax and issues a
deprecation warning to suggest that you upgrade.  We fix the examples in
our cucumber.yml file, but we still have a handful of deprecation warnings
for the old syntax.  Unfortunately these come from hooks defined in
cucumber-rails which is unlikely to be upgraded anytime soon as it targets
older versions of cucumber as well as newer ones, so can't use the new
syntax.