Add rake task to update the organisation of a manual
Recently the Homes and Communities Agency organisation was closed and
are now known as Homes England. There is a manual that exists with the
old, closed, organisation and users that used to be able to access the
manual no longer can. This is because the manual still exists under the
Homes and Communities Agency and doesn't match the user scope.

The following rake task should be ran to update the organisation of the
manual so that users are able to view and edit the document again:

```
bundle exec rake update_manual_organisation[/guidance/capital-funding-guide,homes-england]
```