Fix for change note not showing
For sections where "Minor update" was checked by default, the change note would not appear
when you checked "Major update". This was the expected behavior, but the JS is too ambiguous
and looks for the parent of the change note to make visible, which is unexpectedly an extra
p tag instead of the parent div.

HTML in `app/views/sections/_form.html.erb`:

```
<div class="change_note_container <%= section.minor_update ? 'js-hidden' : nil %>">
  <%= f.text_area :change_note, rows: 20, cols: 40, class: 'form-control short-textarea' %>
  <p class="help-block">This will be publically viewable on GOV.UK.</p>
</div>
```

which gets turned into

```
<div class="js-hidden" style="display: none;">
  <p style="display: block;"><label for="section_change_note">Change note </label><textarea rows="20" cols="40" class="form-control short-textarea" name="section[change_note]" id="section_change_note"></texta
  <p class="help-block">This will be publically viewable on GOV.UK.</p>
</div>
```

Notice that the extra p has the style property set to "display:block" instead of neutralizing
the "display:none" from the div above.

This commit adds a class to the div and points to that instead of using `$("#section_change_note").parent()` which is less reliable.

For: https://govuk.zendesk.com/agent/tickets/2617087