Require full manual section slug when reslugging
Currently when we want to relug a section, we assume
that the manual slug will be a part of the section
slug.

But if the manual slug changes, and the section slugs
do not, then the reslugging implementation wont work,
because we assume that manual-slug + last-part-of-section-slug
will equal the full section slug.

This change the reslugging classes and tasks to require
a full section slug when reslugging.