Let govuk_sidekiq set API headers
Rather than doing it ourselves, we can let the govuk_sidekiq Gem do it
for us. If we don't pass headers as an argument, govuk_sidekiq sets them
for us. More info on how this works can be found here:
https://github.com/alphagov/govuk_sidekiq/blob/31a365c1414be63a801fe032c0397bcb27e1c416/lib/govuk_sidekiq/api_headers.rb#L10-L28

Deleting the params from the method definition and method call means
that we should set the headers manually in the test. But at this point,
we would only be testing that we can read the headers we just set in the
test. This would not really be testing anything of relevance here as
this behaviour is not in the scope of this app, but rather of
govuk_sidekiq. For this reason I deleted the test altogether.

Original commit by Thomas Leese:
https://github.com/alphagov/manuals-publisher/pull/1438/commits/305d0cb78559dd189929f7cc6be337848479d33f