Make cucumber support directory load order independent
In cucumber/cucumber-ruby@2327369 the load order of the support directory
was reversed which mean that world_setup.rb was now being loaded before
all the helper modules were defined. To fix this we move including the
modules into the World into each individual helper file. However since
the helpers are also used in the specs we need to check to see whether
the helper is being used by cucumber or rspec first.