Use database_cleaner-mongoid gem over generic gem
The database_cleaner gem is in the process of splitting out the various
adapters into specific gems which will be complete as of version 2.0, but
in the meantime the generic gem still includes all of the adapters but
due to changes in cucumber-rails it was defaulting to an Active Record
adapter. Fix this by using the specific gem for mongoid and explicitly
stating the orm adapter in the cucumber env.rb file.