Use bulk publishing queue when republishing
Currently if you use the manuals republisher the jobs are placed on the
high priority queue. I think it's unlikely that this is the intention as
this can result it high latency and delay other publishing for
potentially hours.

Instead we can republish using the low priority queue so it continues to
run in the background and gives way to any other publishing in the mean
time.