Fix dubious test broken by linting fixes
These tests broke because there was no longer a call to `is_a?` when
comparing the value of `reportable_object`.

While fixing this I then learnt that this class was using a
SectionEdition object which was pretending to be a Section object. This
class does not have a `manual` method on it and thus fails the test (and
would cause an exception in production).

I've resolved this by removing the call to `Section#manual` and instead
using @report.manual_id directly.