Fix Jasmine issue introduced by installing Rails 6.1 gem
When installing Rails 6.1, and not even running the `app:update`
command yet, the build fails at the `jasmine:ci` step with the
following output:

```
jasmine server started
2021-09-24 08:31:27 +0000 Rack app ("GET /throwFailures=false&failFast=false&random=true" - (127.0.0.1)): #<ArgumentError: wrong number of arguments (given 0, expected 3)>
rake aborted!
Selenium::WebDriver::Error::JavascriptError: javascript error: jsApiReporter is not defined
```

Following the direction in jasmine/jasmine_selenium_runner#6 (comment),
overriding the Jasmine `host` seems to resolve this issue. But
then the tests fail for a different reason, along the lines of
`GOVUK is undefined`. Running `bundle exec jasmine` (without `:ci`)
and viewing in the browser confirms that the application JS does
not appear to be being requested in the Network tab.

I had to change the `jasmine.yml` configuration to pull in a
wildcard of JS/CSS paths. The previous values don't seem to exist
(`assets/application.js` and `assets/application.css` don't seem
to exist anywhere in my repository or output). But running the
`rails assets:precompile` command generates the application JS and
CSS, though it appends a unique filehash to the end of the filename
(before the suffix), meaning it is impossible to specify this
upfront in the YML file without having to change it every time we
make a change to the asset. So I've switched to the `*` wildcard
approach instead. I've also added a `public/` prefix, which is
where the `precompile` command saves its output.

We're not sure why these changes are necessary; perhaps there is
some conditional loading logic in Jasmine which exhibits different
behaviour if Rails 6.1 is detected. More likely, Rails 6.1 changes
how assets are compiled, right out of the box, without being
subject to the `load_defaults 6.0` config option, which seems a
bit of an oversight IMO!

Co-authored-by: Christopher Ashton <christopher.ashton@digital.cabinet-office.gov.uk>