Turn off 'autosave' on Editions association
The ManualRecord model has many Editions. This association was previously configured with `autosave: true`, meaning that changes to Editions would automatically be persisted when calling `save` on a ManualRecord. However that resulted in performance issues due to N+1 queries – eventually resulting in Gateway Timeouts for end users who were attempting to update manuals with ~1000+ editions.

It's not clear exactly _why_ this caused issues considering only the latest Edition would have changed. Our working theory is that it was loading _all_ previous Editions from the database to check whether they've changed and need saving. It could also have been attempting to write each one back to the database, although we noticed that Edition `updated_at` timestamps did not change.

Upon closer inspection we realised that `autosave: true` isn't actually needed in the application – no production code depends on this behaviour. We were able to remove this and gain a huge increase in performance.

An 'after save' callback was added to the `:manual_record` factory used by tests. Without this, tests would fail due to Editions not having been auto-saved when creating ManualRecords.