Remove redundant 'after save' callback
The Edition model had 'after save' and 'after destroy' callbacks which would 'touch' the parent ManualRecord so that its `updated_at` timestamp changed. The purpose of this wasn't entirely clear.

Upon closer inspection, it seemed apparent that this callback was redundant. We don't think it ever ran or had any effect on the ManualRecord.

We also noticed that the ManualRecord's `updated_at` timestamp wasn't even used. To determine when a Manual was last updated, the `updated_at` timestamp of the latest Edition was used. It therefore seemed safe to remove.