Application specific asset path
This sets the asset path to include the application name. It is an
alternative to https://github.com/alphagov/manuals-publisher/pull/2003
which made this something configurable by an env var.

As this will work in all environments set to a consistent pattern we can
avoid the need for this be something that is configured.