Demo of paste-html-to-govspeak feature test
This uses creating-and-editing-a-manual.feature as the place to add the
test due to the existing convention to use this feature file for the
integration tests on creating and editing a manual.

As far as I'm aware there isn't an easy way to do a native copy and
paste with Capybara / Selenium (it would require access to the
clipboard by the headless browser as well as finding something to
select). Therefore I've faked the paste by using Javascript to trigger
an event on the HTML element of type paste.