Add whats new index and controller
What

Added design system layout and CSS
Added whats new index page linked to a yaml for translation
Added back to top for whats-new page
Added type for whats new page
Added updates for whats new page
Extended application js with the components used
Extended application css with the components used and a class for the govspeak text as colour error when importing
Added govuk js as described in design system setup guide
Extended pacakge with front end and dart as described in the setup guide
Extended routes to path whats_new to /whats-new
Added controller for whats-new

Why

Alert users to the features that the consolidation team is adding. Starts as a base for the design system so other pages can be added easily. Emulate as close to the look of Whitehall whats-new page, getting error of colour grey when importing govspeak following the guide so have emulated it in a class