Allow Yarn packages to be installed as part of assets:precompile
This adds an enhancement to `assets:precompile` to restore pre-Rails 7
behaviour that running assets:precompile also runs yarn install. This
allows automatically installing node dependencies.

I've removed the package.json and yarn.lock files from dockerignore so
that yarn assets can be installed in a docker environment.

The motivation for this change is to enable yarn installed packages to
be available for runtime javascript.