Update what's new page "last updated"
Update the "last updated" date on the What's New page to match the most recent (currently the only) update in the list of changes. Technically, the page itself has changed more recently, but it is confusing having the dates be different, and it doesn't seem worth including a what's new update to say that the what's new page itself has appeared/changed.