Migrate the manuals edit page to the design system
Update the edit page to use the "design_system" layout.

This change necessitate updating the layout to enable the `pasteHtmlToGovspeak` functionality, which involves running a javascript snippet on document load to attach a listener to the "paste" event of relevant input fields. The attachment code (which is copied from the legacy layout) uses jquery, which means that jquery needs to be accessible from the design system layout (and to have been loaded earlier than when the general application javascript code is). This attachment code now uses a "data" attribute to tag the input field for listening, rather than the CSS class name used by the legacy layout.

The `govspeak_help` component has been copied from Whitehall (with a few adjustments for Manuals Publisher, for example, removing the text around adding images, which are not supported).