Set default timezone to London
We've had a bug come in that a that was published at 12.01am is saying
that it was published the previous day on the updates section.

Looking at the config I can see that a default timezone hasn't been set.
This means it's using UTC. This doesn't account for BST being an hour
ahead during the summer time so when the updated_at is passed to
publising API it's passing a time of an hour earlier than it was published.

We can resolve this by setting the default timezone to London.

A bulk republish will be required to resolve this issue too.