Move to a supported state machine gem
The state machine gem we had been using had not been updated since 2013, aside from needing to move to a gem that is receving regualar updates we also needed to update the gem as part of our upgrade to ruby 3.2.2 since the old gem was incompatible.

We have moved to a new gem called state_machines to rectify these issues.

Co-Authored-By: Barbara Slawinska <slawinska.barbara@gmail.com>