Disable Rails SASS compression
This resolves a SASS compilation error that has been solved on other projects. The error message is:

Error: "calc(0px)" is not a number for `max'

Disabling sass compression by default and relying on sass style to do it in production should save an unncessary run of the sass compiler and avoid the issue.