Don't attempt to discard draft for published sections
When a manual is published all sections are published as well.
Post publication sections retain a published state expect in 2
circumstances.

1. they are edited
2. they are withdrawn

If they are edited a draft section is pushed downstream to Publishing API
and the Draft Content Store.

When the section is withdrawn, we need to discard the draft from Publishing
API.

There's some logic to handle this in the remove_service which checks to
see if the section is a draft (i.e. has a draft edition in Publishing API)
and if so it attempts to discard the draft.

However, there's an issue with this. Within the remove service assign_attributes
is called on the section. If you go and have a look at the method defined
in the Section model, you can see that it actually creates a new draft
section when called if the latest edition is published. It then passes
change note params to the new draft edition of the section.

This means that when #draft? is called on the section it will always return true.
This causes an issue when the following is called:

```
if section.draft?
  Adapters.publishing.discard_section(section)
end
```

There blows up when a section was in a published state prior to the
remove_service being called as there is no draft section in the Publishing
API database.

There are feature specs which explicity cover that the section is removed
when Published sections are withdrawn. The proposed fix of capturing
the state prior so assign_attributes being called breaks these tests.

I'm pretty sure that the tests are actually testing the incorrect behaviour
which is causing the bug though.

At first glance I thought it might be possible to not bother creating a new
draft edition prior to removing the section from the manual, but it's actually
required for capturing the change note the user inputs so that users can see
why the section was removed.

When an edition is published downstream, any change notes
are pulled from removed sections added as PublicationLogs
These are then rendered on the frontend.